http://bugzilla.liquidsys.com/show_bug.cgi?id=3595 bug fix: 3595 - need a way to tell which messages in CV matched the query
Per conversation betw. Ross and Satish, this is what was decided:

"... Search highlights - this was discussed but never got into the PRD or a bug
report. We have to put that back in but not do it if the search is in:foldername"

So I will turn on highliting for searches that are not "in:<folder_name>"

Copied from Perforce
 Change: 11215