http://bugzilla.liquidsys.com/show_bug.cgi?id=3200 Bug: 3200
Fix change notification handling. The problem was that a view
was changing the model, which messed up the state for
subsequently notified views. Solution is to have the model
(ZmList) listen for changes to the tags that affect its state,
then have the model notify any listening views.

- removed ZmContactCardsView._tagChangeListener()

Copied from Perforce
 Change: 11254