http://bugzilla.liquidsys.com/show_bug.cgi?id=3682 Bug: 3682
After discussions we've decided to always let the MTA generate the delivery errors -- sooo MailServiceException.UNKNOWN_LOCAL_RECIPIENTS is OUT, and MailServiceException.SEND_FAILURE is in.  Modified client so that it looks for the new exception and still prints prettified error message.

Note that if you're testing this in a dev environment, you won't get error checking: this is because dev environments use exch1 by default, and exch1 is configured in a nonstandard way (well, nonstandard for us) -- that is to not reject bad RCPTs right away....

Copied from Perforce
 Change: 11459