For email address objects found in a mail message, add a menu item to go to the associated website.
For example, the address "bob@yahoo.com" will result in
a menu item "Go to www.yahoo.com". Only the last two
parts of the domain are considered, so an address of
"fred@xml.apache.org" or "wilma@mail01.apache.org"
will produce a menu item "Go to www.apache.org".

When clicked, the website opens in a new window.

Copied from Perforce
 Change: 12465