Async: replenishment and pagination
Plus miscellaneous cleanup, reworking, and fixes.

Views now know who their controller is. That was 
necessary for async replenishment, which now happens
in response to a notification rather than a SOAP response.
The change listener in ZmListView gets the notification
and figures out that it needs to grab more items, which
it does with an async search.

Now that each view has a handle to its controller, I'll
take a look at getting rid of getControllerForView(). Yay!

Async pagination called for the first instance of a
response callback that gets overridden. Several controllers
were overriding _paginateCallback in the same way; I've
moved it to ZmMailListController and renamed it.

The replenishment and pagination routines used to pass a
callback to search() so that it didn't just display the entire
result set. Now that's just a flag (noRender), and the
async callback subsumes the earlier one.

Removed callbacks from item actions; the intent is that 
they be "leaf" actions and the UI gets updated based on
notifications. The one exception is moving convs, since the
notifications only tell us which messages moved, and we
can't connect those to the moved conv.

Removed several try/catch blocks that were rendered
useless by conversion to async.

Removed ZmCalTreeController._deleteListener, it did
what its parent class did.

Copied from Perforce
 Change: 14498