Clean up how the overview is set, especially for calendar.
Now, when we switch to a view that belongs to a different
app (even if it's not a top-level view), we change the current
app toolbar and the overview to reflect the app change. So if
you're in your Inbox and select "New Appointment", you get
switched to the calendar app, including the display of the
mini-calendar.

- Changed calendar app to use numeric IDs for its views rather
  than strings
- Each app has a set of trees it shows in the overview panel

Copied from Perforce
 Change: 15039