http://bugzilla.liquidsys.com/show_bug.cgi?id=4434 bug: 4434 - Calendar drop-down should only contain calendars in which user has write access
Client now shows only those calendars which user has write access. The implementation of this code will change once we get full server support post Birdseye. 

The way it works today, when the client first receives a refresh block we process all the folders, then attempt to get the permissions of those folders which are "links" (via GetFolderRequest). The response to this request returns the "perm" values for each folder (aka ZmCalendar) which is artificially stored in their respective shares array. 

The non-hacky way to do this would be to have the server always send us permissions for all folders w/in the refresh block. As Dan mentioned in the previous comment, this will happen post Birdseye.

Should we keep this bug open so we remember to fix this hack?

Copied from Perforce
 Change: 16054