http://bugzilla.liquidsys.com/show_bug.cgi?id=4356 bug: 4356 - Compose panel does not get dismissed when editing invite replys
So the way it works now is, we check in handleDelete if the current view is compose and if so, check if the number of msgs in the conversation is only one and that it is an invite - then and only then, we pop prematurely. 

The second pop happens normally (via _processSendMsg in ZmComposeController) which really pops the CV -- its a little backwards but theres no other way besides changing the way we handle notifications.

I also took out listNotify call in _handleResponseSendInviteReply() in ZmMailMsg since modify notifications after sending the invite does the same thing (so we dont call handleDelete twice)

NOTE: This fix will not be integrated into Birdseye for the sake of avoiding any regressions.

Copied from Perforce
 Change: 16437