http://bugzilla.liquidsys.com/show_bug.cgi?id=4484 Bug: 4484
Handle auth expired, so that the client picks up where
it left off after the user logs in again.

sendRequest() now takes an optional "execFrame" arg,
which is a callback that will be executed after a re-login
if an auth expired exception occurs. For async calls, that
replaces the call to _handleException() with a catch block.

The following actions will continue after a re-login:
- fetch msg (inc invites)
- send msg
- item action
- organizer action
- search

There may be more we'll want to set up, but those are
the major ones.

Copied from Perforce
 Change: 16888