http://bugzilla.zimbra.com/show_bug.cgi?id=2203 bug: 2203 - Improper tooltip for date of first message in list
This was happening b/c we use the user's localtime to compute the elapsed time since the message arrived. It just so happens the date for the new message as returned by the server is ahead of the user's local time (message arrived in the future).

We now ignore setting the elapsed time if the delta between the time the message arrived and the local time is negative :)

Copied from Perforce
 Change: 18570