http://bugzilla.zimbra.com/show_bug.cgi?id=3700 Bug: 3700
Resolve a class of bugs where a search that is based on
an organizer (eg "tag:foo") is done, and then a change is
made to the organizer that causes any re-use of the search
(pagination, sorting, opening a conv) to fail.

Certain folder and tag actions now trigger a listener that
looks to see if the affected folder or tag is part of the current
search.

If a folder or tag is renamed, or a folder is moved, we update
the query string, and everything works fine. If a folder is
deleted, the results are blanked - no change there.

If a tag is deleted, we leave the results up (the tag column is
updated). Pagination and sorting are disabled. If we're in CLV,
then double-clicking the conv will open the conv, but there won't
be any highlighting in the CV's message list.

- added enableSorting() to DwtListView
- removed searchString from various places, it's available through
  the current search
- added a routine to escape strings for use in regexes

Copied from Perforce
 Change: 19310