http://bugzilla.zimbra.com/show_bug.cgi?id=5100 bug: 5100 - some messages show blank Reading Pane in Firefox
I was not able to repro this bug for all testcases posted for this bug. However, a couple did lead to the following error message (via JS console in FF):

----------
Error: Node cannot be inserted at the specified point in the hierarchy
Source file: http://localhost:7070/zimbra/js/zimbraMail/mail/view/ZmMailMsgView.js?v=060112141556
Line: 440
----------

Digging further shows we were trying to append an anchor node to a textnode which threw an exception (which we never caught) - hence the "blank" messages.

The fix here is to catch the exception and move along in the DOM processing. This may or may not be the correct fix (since we could potentially lose some content). 

Therefore, reassigning bug to Mihai for a closer look (re: review this fix) as well as find out why we end up trying to append an illegal node to begin with.

Also, bug 5345 is said to be a duplicate of this bug, however with the given fix, it still causes IE to CRASH. Therefore, 5345 needs more research - also asking Mihai to take a look.

This bug fix will be integrated into mainline once Mihai has had a chance to review the fix.

Copied from Perforce
 Change: 19508