http://bugzilla.zimbra.com/show_bug.cgi?id=4681 bug: 4681 - "Reply all" doesn't fill in CC/BCC fields when opening in new window
Found a couple of articles on the web describing what we're seeing with this bug:

http://www.thescripts.com/forum/thread91190.html
http://english.kakeboksen.org/archives/000308.html

Apparently, opening a new window creates a new context (nothing new here) which means we lose all type information for objects passed into a child window from the parent.

The only decent fix for this (where both IE and FF are happy) is to preform a deep copy of the ZmMailMsg that is being used to "reply all" -- A better solution would be to figure out how to maintain the passed in object's integrity ... for now, this is a suitable fix.

Copied from Perforce
 Change: 19525