http://bugzilla.zimbra.com/show_bug.cgi?id=290 bug: 290 - Script error when loading a huge set of contacts
We now process 500 contacts at a time and then pause for 250ms for the UI to be a little more responsive to user action. IE is still sluggish, but neither browsers should show the "Script on this page..." error (on initial load).

Importing contacts is another issue altogether (please write up a new bug report if need be).

Copied from Perforce
 Change: 19604