http://bugzilla.zimbra.com/show_bug.cgi?id=5569 Bug: 5569
Make autocomplete more usable in IE with a large
contact list.

Tested against a collection of ~3900 contacts. Before
these changes, IE would take around 30 seconds to
do initial autocomplete matching. In Firefox, it only took
1.5 seconds to do that and show the list.

The solution is to do pre-matching as contacts are loaded.
A matching structure is created for the first letter of the
contact's first name, last name, and each email address.
That's also done for the first two letters.

The contacts are loaded in chunks, with pauses to give the
browser some CPU. During the load, the browser will be
less responsive. There are a couple constants that can be
tuned to adjust that.

If autocomplete is attempted before the contacts are loaded,
a warning will appear in the status box, and no list is shown.

Loading the 3900 contacts with the pre-matching takes 15
seconds in Firefox, and 7 minutes in IE. After loading is
complete, autocomplete is quite responsive.

Copied from Perforce
 Change: 20404