http://bugzilla.zimbra.com/show_bug.cgi?id=5569 bug: 5569 - Revert fix for this bug: change #20404 - this fix was causing Jim's account to completely freeze on login for both browsers.
Something related to Jim's contacts was causing this code to run infinitely but b/c there are timed actions, browser never throws up the infinite loop dialog (instead it pegs the CPU).

Instead of trying to find the bug for this fix, we decided to back it out completely since there are other minor issues that should be resolved (i.e. not optimize autocomplete for FF - as requested by Satish).

Copied from Perforce
 Change: 20539