http://bugzilla.zimbra.com/show_bug.cgi?id=5086 bug: 5086 - Reply & Forward from CLV has blank msg. content
Why this bug was happening:
- User receives an HTML message, and opens it from the CLV - this makes a request to the server to retrieve the first "hit" message in that conversation - and additionally, asks for the HTML version if it exists.

- Server returns the HTML version, but the user's settings for composing a new message is set to Text. 

That means client has to convert the HTML version to text. 

There were two bugs that were causing an "empty" message:
1. Client was asking for the text part of a message which had only retrieved the HTML part
2. Client was attempting to parse an HTML document (so it can then convert it to text) but was assuming that HTML document was not fully formed (doesnt have HTML, HEAD, TITLE tags). 

The parsing routine needs a root element to begin parsing out the DOM objects of an HTML message, but creating one by wrapping the content (which in this case, contained the fully formed HTML) around a DIV only returned the TITLE element.

The solution here is to create a temp DwtIframe that will take the HTML content and automatically build valid DOM tree for us - which we then can use to pass into the parsing routine.

Copied from Perforce
 Change: 20697