http://bugzilla.zimbra.com/show_bug.cgi?id=5603 bug: 5603 - "Could not send message: $0" error if message size is more than max allowed
We now show the following error message when attempting to send a message larger than the max allowed: 

"Could not send message: mail.SEND_FAILURE"

The client cannot give any more detail about this error message b/c that is all the server sends. Clicking on the down arrow (details) button will give more info on why the send failed and shoud be sufficient.

The real fix then is to not sure the {0} and actually display the correct error code.

Copied from Perforce
 Change: 20724