http://bugzilla.zimbra.com/show_bug.cgi?id=6518 bug: 6518 - Cannot share calendar due to Parse error
Fixed the HTML "parse error" Safari is complaining about... 

Unfortunately, for some reason, the receiver of the invite gets a blank content. The content going to the server seems to be populated correctly (in comparison to the same share request made by FF) - may need some server help to find out why the content is blank (need to make sure server is receiving the right data to set the content in the message containing the share request).

Copied from Perforce
 Change: 23077