http://bugzilla.zimbra.com/show_bug.cgi?id=5394 Bug: 5394
Adjust an appt's start or end based on changes to the other one. If the user changes
the start time, change the end time so that the appt duration is maintained. If the
user changes the end time, we leave things alone unless the end time is changed to
before the start time, in which case we change the start time to maintain the appt
duration.

Copied from Perforce
 Change: 23306