http://bugzilla.zimbra.com/show_bug.cgi?id=6788 bug: 6788 - Can use HTML format only once using Safari
Safari is weird. It loses the document when you switch out of HTML mode and into TEXT. So everytime you switch back into HTML mode, you have to recreate the document.

The shrunken compose area, when in TEXT mode, was caused by trying to calc. the height of the area by subtracting heights of widgets above it incl. the non-existent HTML toolbars (which were returning undefined in Safari and resulting in a "NaN" instead of zero).

Copied from Perforce
 Change: 23552