http://bugzilla.zimbra.com/show_bug.cgi?id=6821 bug: 6821 - Signature problems when HTML and Open in New Window set
Interesting bug. 

With HTML compose turned on as default and composing in new window, the client assumes the HTML editor widget is fully initialized when setting up the initial "body" of the message. Apparently there is a timing issue, when opening a new window, between setting the content and then resetting it w/ the signature. Therefore, the content was never getting set (or signature depending on your combination of options set) and one or the other would get blown away.

This has been fixed for cases where there is initial content when composing (in new window or not).

Copied from Perforce
 Change: 23621