http://bugzilla.zimbra.com/show_bug.cgi?id=6955 bug: 6955 - Mail>Calendar: Appt's contents are lost
This was happening if user had "View mail as HTML" option set and composed an appointment as text. When the user tries to open the appointment, client makes a GetMsgRequest and asks for HTML which the server is able to return (since creating appointments sends both text and HTML parts). Since no text part is returned in the GetMsgResponse, we were failing to convert the HTML part to text. Good catch.

Copied from Perforce
 Change: 23827