http://bugzilla.zimbra.com/show_bug.cgi?id=7016 bug: 7016 - Previously saved draft edited in new window does not save modifications
Actually, this bug can be more easily repro'd by:
- open compose, write something 
- save draft, close compose
- reopen same draft, write something more
- resave draft, close compose
- reopen draft (doing so will show old draft content)

In other words, doesnt really have anything to do with opening a new window.

This was happening because we cache messages that are fetched from the server. When reopening the draft we pull it from the cache which does not have the changed content from the last time it was saved. We now no longer cache draft messages for this reason.

Copied from Perforce
 Change: 23989