http://bugzilla.zimbra.com/show_bug.cgi?id=5182 Bug: 5182
Handle errors when creating a calendar folder
sync'ed to a remote calendar.

- In ZmZimbraMail, the async response handling has
been changed to handle notifications before running
the response callback. That is necessary for a chain 
of callbacks which have dependent notifications.

- Rewrote calendar folder creation to work like
other organizer creation. Creating a calendar folder
is now an instance method of ZmCalendar. The 
content of the new calendar dialog is now created
via innerHTML rather than DOM methods, and it's
now a ZmDialog.

Copied from Perforce
 Change: 24495