http://bugzilla.zimbra.com/show_bug.cgi?id=5964 Bug: 5964
Fix relogin when there's no exec frame.

There were several problems. One is that we
were recreating components during a relogin-triggered
startup, which led to the odd double-printing of the
current app.

Another is that we were creating the overview fresh
during startup, which in relogin resulted in the overview
getting blanked.

We were also sending the user to the default app (mail)
after relogin. Now we keep them in the app they were in.

We now explicitly clear the session ID during relogin so
that the refresh block gets processed.

Copied from Perforce
 Change: 24839