http://bugzilla.zimbra.com/show_bug.cgi?id=7526 bug: 7526
Replace the read-only appt dialog (which showed
appts for which you are not the organizer) with an
appt view. The view is similar to the single msg view,
with appt notes at the bottom and everything else
in a header section at the top.

- Uses the standard calendar toolbar, so you can
switch to a calendar view, delete the appt, etc
- Button for the current view is now disabled

Copied from Perforce
 Change: 25581