http://bugzilla.zimbra.com/show_bug.cgi?id=8468 Bug: 8468
Refine GAL autocomplete timeout handling.

As before, timeout for GAL autocomplete is
500 ms, and 5 timeouts in a row turns off the
feature. Now there's a new setting which is
bound to the session. If it gets turned off, a new
setting will show up in Options on the Address Book
tab which asks if you want to re-enable GAL
autocomplete.

Copied from Perforce
 Change: 29343