http://bugzilla.zimbra.com/show_bug.cgi?id=10015 http://bugzilla.zimbra.com/show_bug.cgi?id=7896 Bug: 10015, 7896
Revamp GAL autocomplete.

Local results (from the user's address book) are
shown immediately. If GAL autocomplete is enabled,
the call is issued and the user will see "Waiting for GAL..."
at the bottom of the list. When the call returns, the
"waiting" message is removed and results are merged
into the list.

If no results are returned, the "waiting" message will
appear for a minimum of 1 second, to prevent a 
"flashing" effect.

There is still a timeout mechanism, but the timeout
has been increased from 0.5 to 15 seconds since
local results no longer wait for GAL results. As before,
5 consecutive timeouts temporarily disables the
feature.

Each entry has an icon to its left indicating whether
it's local or GAL. If the user changes whether GAL
autocomplete is enabled, cached results are cleared.

These changes apply to all clients of GAL autocomplete:
compose mail, new appt, etc.

Copied from Perforce
 Change: 31752