http://bugzilla.zimbra.com/show_bug.cgi?id=9838 Bug: 9838
Pull keyboard shortcuts out of code into
properties files so they can be i18n'ed.

The shortcuts are pulled from the properties
files into a JS class by a servlet in the same 
way that strings are. During startup, they are
loaded from the JS class into a map in the
form that the keyboard mgmt code needs.

TODO:
- platform-sensitive shortcuts
- move tab view shortcuts to DwtTabView
- add back debug shortcuts
- come up with better way for servlet
to figure out basename

Copied from Perforce
 Change: 32626