http://bugzilla.zimbra.com/show_bug.cgi?id=12625 Bug: 12625
Fix bug where shortcuts stop working.

The problem was that we push a default
shortcuts handler (normally ZmZimbraMail)
when a dialog pops up, and pop it when the
dialog pops down. That's so that the dialog
can handle keys independent of focus. We
disable keyboard nav while a SOAP request is
in transit, and if we tried to pop the dialog's
default handler during that time, it failed due
to a check. Fix was to remove the check.

Copied from Perforce
 Change: 37510