http://bugzilla.zimbra.com/show_bug.cgi?id=9525 Integrate 42631 from main to FRANK
-----------------

Bug: 9525

Fix "ZmController is not defined" bug.

The msg controller created by the new window
for viewing the RFC822 adds a listener to the tag
list, which remained after the window (and the
controller that provided the context for the listener)
was destroyed. Any change in the parent window to
the tag tree would then try to invoke that listener
and fail.

The fix is to remove the listener on unload of the new
	window.

Copied from Perforce
 Change: 42632