http://bugzilla.zimbra.com/show_bug.cgi?id=14385 Bug: 14385
Defer notifications for apps that are not yet loaded.

Each app now registers what its main package is. If
we get notifications and the app has special handling,
we defer them if that package is not loaded. The app
sets up a post-load function that handles them after
the package gets loaded.

Copied from Perforce
 Change: 43532