http://bugzilla.zimbra.com/show_bug.cgi?id=10325 Bug: 10325
Began the process of pulling app-specific operations into
the apps, and redid a lot of overview and tree controller
stuff.

Apps can register a list of operations that they handle. So far,
this is just the NEW_* operations, but that will be expanded.
The app defines handleOp() for the ops it supports.

_inNewWindow() moved from ZmListController to ZmApp.

ZmController::_showDialog became static, and takes an actual
AjxCallback rather than the args to create one.

Removed "loc" from arg list for ZmDialog::popup, it was never
used. Also applied to its subclasses.

Got rid of the hokey overview IDs that had an app component
(eg "ZmZimbraMail-mail"). Those were for handling the different
versions of the Searches tree view, which only shows saved
searches relevant to the app. Instead, just reset that tree view
when the overview is set.

The top-level functions for rendering a tree view now take a
params hash rather than a bunch of args.

Came up with a generic way to move tree items between different
tree views, so that I could remove all the crappy code that was
handling moving searches between the Folders and Searches trees.
Added getTreeItemById() to ZmOverviewController, which will
search multiple trees for the item. ZmSearchTreeController got
a lot smaller.

Added a "handled" property to ZmEvent (change events), so that
an event won't get handled more than once.

ZmTreeController now calls _postSetup() if a tree view was
created, which eliminates the need for its subclasses to override
show() so that they can fix up the tree view. It handles checkboxes
and colors by default.

- DwtControl::reparent() now takes an index at which to add the
  item when it's reparented. Not used yet, but potentially
  useful.

- DwtDialog::registerCallback() can take either an AjxCallback
  or the args to create one. Makes it possible to pass an
  existing callback.

Copied from Perforce
 Change: 44033