http://bugzilla.zimbra.com/show_bug.cgi?id=14623 Bug: 14623
Reorganized how and when the view button and
menu are updated.

Before, ZmCurrentAppToolBar.showViewMenu()
would get called from 2 to 4 times on each view
switch. That's now down to once.

Removed ZmListController._setViewMenu(), which
was redundant. The view menu is set when a view
is switched and is called by ZmZimbraMail.setActiveApp().

An app controller can set up associated sub-views if it
manages its own views (eg ZmCalViewController). Also,
if convs are disabled we now have the view button show
"Reading Pane On" or "Reading Pane Off". That's done by
registering a callback for the view.

Copied from Perforce
 Change: 44382