http://bugzilla.zimbra.com/show_bug.cgi?id=14831 Bug: 14831
Fix to prevent duplicate folders in the overview.

Folder creation is deferred during startup. The problem
was that the "inStartup" flag wasn't being set when we
called startup() as a result of a re-login, so the folder would
be created as usual on the receipt of a refresh block, and
then the deferred version would be created when we went
to the app.

Copied from Perforce
 Change: 47677