http://bugzilla.zimbra.com/show_bug.cgi?id=15638 Bug: 15638
Hybrid mail view (alpha)

The hybrid mail view makes convs with more than one msg
expandable, and adds a reading pane. The first msg in the
selected conv will appear in the reading pane. Expanding a
conv (via a + icon) inserts rows for its msgs below it.

Three columns have been added to the headers for CLV to
support msg display: the expansion icon, msg status, and
folder. The size column shows either number of msgs (conv)
or size (msg).

We need to visually distinguish msgs from convs. So far, I give
msg rows a background of light blue (that should be moved to
skins), and indent the text in the subject column (which is actually
the msg fragment). Also, only msgs will have something in the status
or folder column.

If a conv has more than one page of msgs, the last msg will
become expandable (get a + icon), and expanding it loads the
next page of msgs.

Lots of changes, many because we now have a list view
with two different types of underlying data (conv and msg).

Features:
- "Hybrid" shows up as option in view menu; VH is shortcut
- Enter and O open a conv in CLV; here they expand/collapse it
- Dbl-click also expands/collapses a conv
- The actions above also apply to an expandable msg
- If reading pane is off, dbl-click on a conv opens it into CV, and
  dbl-click on a msg opens it into MV

Changes:
- Added DwtListView::_getRowIndex which returns index of row
  for given item (may differ from index of item in list)
- Added way to prevent the same "modified" notification from being
  handled more than once by change listeners. Ran into this because
  a msg notif tells its containing conv, then we get a notif for that conv.
- Rearranged the "group mail by" stuff; factored some of the constants
  down into relevant controllers.
- Factored _defaultView() up into ZmListController
- Factored switchView() up into ZmMailListController (reading pane is
  handled by ZmDoublePaneController)
- Factored _setupViewMenu() and getTitle() up
- Made fetching the first msg optional on conv load since we don't want
  to do that when expanding a conv
- _msgListView -> _mailListView in ZmDoublePaneView and subclasses
- Made view names shorter (ZmController) since they're used in
  tons of element IDs
- Actions on a hybrid list split the list into two lists, one of convs and one
  of msgs, and act on those separately. Actions in ZmListController look
  at the first item to get the list object, and then at the view's underlying list.
- Views can declare multiple types of events to support change events for,
  rather than just checking for their main type

Todo:
- Handle create notifications
- Add hybrid option to LDAP "groupMailBy" attribute
- Preferences support
- More testing
- Cleanup

Copied from Perforce
 Change: 48616