http://bugzilla.zimbra.com/show_bug.cgi?id=16207 Bug: 16207
Implement zippy lists. When mail is displayed as a list of items
with a reading pane below, wait for a small pause before loading
a msg into the reading pane. That way you can hit the arrow keys
rapidly without having to wait for each successive msg to load.

Copied from Perforce
 Change: 49904