http://bugzilla.zimbra.com/show_bug.cgi?id=13764 http://bugzilla.zimbra.com/show_bug.cgi?id=16354 http://bugzilla.zimbra.com/show_bug.cgi?id=16468 Bug: 13764, 16354, 16468
(forgot checkin comment)

Refactored _createItemHtml() and the functions it calls.
There is now a useable default implementation in
DwtListView, which does the following:

	- create a DIV and set its style class
	- create a TABLE
	- create a TR, set its ID and class
	- create a TD and set its content for each column

Each of the above tasks is contained within a function
that subclasses may override. In general, each function
(including _createItemHtml) takes two arguments: the item,
and an optional hash of params which includes options and
other data. Some of these functions were moved up from
ZmListView to DwtListView, others are new.

The most common case for a subclass is to override just
_getField() to set the content of the cells. In general,
a lot of list view definitions of _createItemHtml got
replaced by an override of _getField.

It's easier to piggy-back on other views' implementations
of list rows, for example the hybrid and mixed views.

Got rid of FIELD_PREFIX and PREFIX_FIELD for mapping
between field IDs and their prefixes (for use in DOM IDs).
Now, the field ID is used as the prefix. The ID must be
two letters in length. They are defined in ZmItem rather
than in ZmListView.

- added separator to column header IDs;
  see DwtHeaderListItem.getHeaderField()
- created new class ZmContactsHelper for common search code
  shared by ZmContactPicker and ZmGroupView
- move some style modifiers to DwtCssStyle
- fix button state setting in DwtChooser
- fix search type for adding contact group
- centralized column identifiers under ZmItem.F_*
- changed name of style 'Row-Focus' to 'Row-focused'
- decreased list selection wait pause from 500 to 300ms
- changed email addr and header name IDs from int to string

Copied from Perforce
 Change: 51355