http://bugzilla.zimbra.com/show_bug.cgi?id=16944 http://bugzilla.zimbra.com/show_bug.cgi?id=16948 Bug: 16944, 16948
Reworked overview management of tree views, so that
a tree view does not remain clipped from the DOM.

When the overview is set, existing views are clipped (so
that we can set order for new list of tree views). Then tree
views are created/re-added in the proper order. Finally, any
tree views that were clipped and not re-added are added back
and hidden.

Copied from Perforce
 Change: 53221