Bug: 15810, 16995
Handle single and double clicks in double-pane view.

The first part of the fix was to not use the busy overlay
during SOAP requests triggered by single clicks, so that
the second click would be processed. Although that solved
the immediate problem, we were still issuing extra SOAP
requests. For example, double-clicking on a msg in msg mode
resulted in 3 GetMsgRequest's (click, click, dbl-click).

The solution is to keep track of pending msg requests, and
not duplicate them. To handle dbl-click, we override the
response's callback in ZmMailMsg, so that the GetMsgRequest
from the first single click event gets handled as if it were the
dbl-click event.

http://bugzilla.zimbra.com/show_bug.cgi?id=15810
http://bugzilla.zimbra.com/show_bug.cgi?id=16995

Copied from Perforce
 Change: 58078