Bug: 14815
Optional upsell content for apps.

An app can be disabled but have upsell turned on, along
with an upsell URL. If that's the case, it will have a button
in the app chooser. Pressing the button will display the
content from the upsell URL in an iframe in the content
pane.

Briefcase is set up as an example. To try it, go to ZmSettings
and set BRIEFCASE_UPSELL_ENABLED to true.

I'm currently guessing that we'd have COS settings along
the lines of:

zimbraFeatureBriefcaseUpsellEnabled
zimbraFeatureBriefcaseUpsellUrl

for each app for which we want to support upsell.

http://bugzilla.zimbra.com/show_bug.cgi?id=14815

Copied from Perforce
 Change: 58383