Bug: 17734
Error handling for new login page.

Changed the way we get error messages for server
exceptions so that we just use the error code as the
key (like the standard client does). That way they're
easily accessible via a resource bundle.

- Removed ZmCsfeException.define(), replaced it
with direct definition of exception codes

http://bugzilla.zimbra.com/show_bug.cgi?id=17734

Copied from Perforce
 Change: 58399