Bug: none
Make sure operation listeners (such as _replyListener) are first
in the list of menu item listeners.

Before, menu items had two selection listeners:
- one to pop down the menu, which in turn invokes a popdown
  listener that clears right-click selection via a timed action
- one to perform the appropriate action (REPLY, etc)

The timed action was needed because otherwise the right-click
selection is cleared before the action has a chance to see what it
is. That was a bit hacky and doesn't work during debugging. The
cleaner way is to run the action and then pop down the menu
and clear the right-click selection.

To accomplish this, an index is passed along to addListener().

Copied from Perforce
 Change: 64005