Bug: 20888
Handle mail-related preferences changes by updating the UI.
The following mail prefs are handled dynamically:
- page size
- group by (msg or conv)
- reading pane
- show fragments
- display msg as HTML or text

In addition to listening for changes to a particular
setting, you can listen for changes to the settings as
a whole. That was necessary so we don't re-run the
current search more than once if multiple mail prefs
are changed.

http://bugzilla.zimbra.com/show_bug.cgi?id=20888

Copied from Perforce
 Change: 71049