bug: 22480 It seems that the choice of an INPUT TYPE=SUBMIT (as opposed to an A tag) for the button is forcing some padding on either side of the button -- a partial fix was accomplished by adding:   text-align:left to the classes   .TbTop INPUT, .TbBottom INPUT {...}
There is still a lot of space after each button (between the button and the next) but at least the text snugs up close to the icon.  I do not know of a CSS way to remove this space.

Suggested real fix is to use <A> tags instead of INPUT TYPE=SUBMIT tags for the toolbar buttons.  This should be done by someone more familiar with the HTML code, if it is even possible at all.

http://bugzilla.zimbra.com/show_bug.cgi?id=22480

Copied from Perforce
 Change: 76896