Bug: 23526 - refresh block with large number of folders/searches is slow
Handle situation where a large number of tree items need to be
created and shown, for example if someone has hundreds of
top-level folders.

If there are more than 100 tree items to show at once, we show the
first 100 and then add a special tree item named "Show remaining folders".
If the user clicks on that, we load the remaining folders 50 at a time (so
that the browser doesn't get overloaded), via a timed action.

Works for overview and move dialog.

http://bugzilla.zimbra.com/show_bug.cgi?id=23526

Copied from Perforce
 Change: 91295