Bug: 26305 - Advanced search attachment + domain scroll bars + busy/spin
Domain picker enhancements.

Instead of loading the full set of domains up front, work
via domain searching. When the picker comes up, the top
20 domains are shown. There is now a search box that
accepts search text. Text will be searched for at the beginning
of a domain unless it has a dot (.) in it, in which case we
search for it at the end of domains. A search is done once
there's a 300ms pause in typing. Up to 20 results are shown.

http://bugzilla.zimbra.com/show_bug.cgi?id=26305

Copied from Perforce
 Change: 95565