bug: 27519
other than Email tab scroll bars is not visible. Need to trim down the toolbar to avoid scroll bar in Email. 

we can have pagination like 50/250 or just next/previous arrows instead of 1-50 of 250 and remove the Move and Go button this will make the Email tab with no scroll in 1024 x 768

http://bugzilla.zimbra.com/show_bug.cgi?id=27519

Copied from Perforce
 Change: 96450