bug: 27148
Made sure the view components are re-fitted to perfection after setting all the components. This works fine in FF and only fixed for IE. But, thought this is a perfect way to go and might help even FF when many additional components with varying sizes/pos's are added.

http://bugzilla.zimbra.com/show_bug.cgi?id=27148

Copied from Perforce
 Change: 96572