Bug: 28625 - Error when creating a new contact
There's a special fake folder created for My Card, with
ID -15. Normally when editing a contact there's a select
for choosing the target address book, but that's hidden
for PV. The My Card folder was first in the hidden select,
so it was getting used as the target address book.

Later, the folder list is sorted and that folder is no
longer first, so adding a contact works on the second and
subsequent tries.

http://bugzilla.zimbra.com/show_bug.cgi?id=28625

Copied from Perforce
 Change: 100999